requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
sqlparse==0.5.0
tzdata==2024.1

